app/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.2
filelock==3.16.1
fsspec==2024.9.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.1
mpmath==1.3.0
networkx==3.4.1
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
python-multipart==0.0.12
requests==2.32.3
sniffio==1.3.1
starlette==0.39.2
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
